refactor: move translation related functions to `shared`
