+++ shared/lib/error-utils.js
@@ -1,8 +1,6 @@
 import { memoize, escape as lodashEscape } from 'lodash';
-// TODO: Remove restricted import
-// eslint-disable-next-line import/no-restricted-paths
-import getFirstPreferredLangCode from '../../app/scripts/lib/get-first-preferred-lang-code';
 import { fetchLocale, loadRelativeTimeFormatLocaleData } from '../modules/i18n';
+import getFirstPreferredLangCode from './get-first-preferred-lang-code';
 import { switchDirectionForPreferredLocale } from './switch-direction';
 
 const defaultLocale = 'en';
