+++ shared/lib/translate.test.ts
@@ -1,9 +1,5 @@
-import {
-  getMessage,
-  fetchLocale,
-  FALLBACK_LOCALE,
-} from '../../shared/modules/i18n';
-import { t, updateCurrentLocale } from './translate';
+import { getMessage, fetchLocale, FALLBACK_LOCALE } from '../modules/i18n';
+import { t, updateCurrentLocale } from '../../shared/lib/translate';
 
 const localeCodeMock = 'te';
 const keyMock = 'testKey';
@@ -12,8 +8,8 @@ const messageMock = 'testMessage';
 const messageMock2 = 'testMessage2';
 const alternateLocaleDataMock = { [keyMock]: { message: messageMock2 } };
 
-jest.mock('../../shared/modules/i18n');
-jest.mock('../_locales/en/messages.json', () => ({
+jest.mock('../modules/i18n');
+jest.mock('../../app/_locales/en/messages.json', () => ({
   [keyMock]: { message: messageMock },
 }));
