+++ ui/pages/notifications/notification-components/erc721-sent-received/erc721-sent-received.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
 import { NotificationServicesController } from '@metamask/notification-services-controller';
 import { CHAIN_IDS } from '../../../../../shared/constants/network';
-// TODO: Remove restricted import
-// eslint-disable-next-line import/no-restricted-paths
-import { t } from '../../../../../app/scripts/translate';
+import { t } from '../../../../../shared/lib/translate';
 
 import { type ExtractedNotification, isOfTypeNodeGuard } from '../node-guard';
 import {
