+++ ui/pages/notifications/notification-components/lido-withdrawal-requested/lido-withdrawal-requested.tsx
@@ -23,9 +23,7 @@ import {
   getNetworkDetailsByChainId,
   getUsdAmount,
 } from '../../../../helpers/utils/notification.util';
-// TODO: Remove restricted import
-// eslint-disable-next-line import/no-restricted-paths
-import { t } from '../../../../../app/scripts/translate';
+import { t } from '../../../../../shared/lib/translate';
 import {
   TextVariant,
   BackgroundColor,
