+++ ui/pages/notifications/notification-components/swap-completed/swap-completed.tsx
@@ -6,9 +6,7 @@ import {
   type NotificationComponent,
 } from '../types/notifications/notifications';
 import { CHAIN_IDS } from '../../../../../shared/constants/network';
-// TODO: Remove restricted import
-// eslint-disable-next-line import/no-restricted-paths
-import { t } from '../../../../../app/scripts/translate';
+import { t } from '../../../../../shared/lib/translate';
 
 import {
   NotificationListItem,
